+++ packages/mongo-nav/src/MongoNav.story.tsx
@@ -4,7 +4,7 @@ import { select, boolean, text } from '@storybook/addon-knobs';
 import { css } from '@leafygreen-ui/emotion';
 import MongoNav from '.';
 import LeafygreenProvider from '@leafygreen-ui/leafygreen-provider';
-import { Product, ActiveNavElement } from './types';
+import { Product, Platform, ActiveNavElement, Mode } from './types';
 
 const setStorybookWidth = css`
   width: 100%;
@@ -15,7 +15,12 @@ storiesOf('MongoNav', module).add('Default', () => (
   <LeafygreenProvider>
     <div className={setStorybookWidth}>
       <MongoNav
-        mode="dev"
+        mode={select('mode', Object.values(Mode), Mode.Dev)}
+        activePlatform={select(
+          'activePlatform',
+          Object.values(Platform),
+          Platform.University,
+        )}
         activeProduct={select(
           'activeProduct',
           Object.values(Product),