+++ packages/mongo-nav/src/user-menu/UserMenu.spec.tsx
@@ -8,6 +8,7 @@ const account = {
   firstName: 'Leafy',
   lastName: 'Green',
   email: 'leafy@mongodb.com',
+  admin: false,
 };
 const onLogout = jest.fn();
 const onProductChange = jest.fn();
@@ -34,9 +35,9 @@ describe('packages/mongo-nav/user-menu', () => {
     expect(trigger).toBeInTheDocument();
   });
 
-  test('when "activeProduct" is set to "cloud, renders Cloud MenuItems', () => {
+  test('when "activePlatform" is set to "cloud, renders Cloud MenuItems', () => {
     const { getByTestId, getByText } = renderUserMenu({
-      activeProduct: 'cloud',
+      activePlatform: 'cloud',
     });
     const trigger = getByTestId('user-menu-trigger');
     fireEvent.click(trigger);
@@ -54,7 +55,7 @@ describe('packages/mongo-nav/user-menu', () => {
 
   test('renders university MenuItems when university dropdown is clicked and closes other SubMenus', () => {
     const { getByTestId } = renderUserMenu({
-      activeProduct: 'cloud',
+      activePlatform: 'cloud',
     });
     const trigger = getByTestId('user-menu-trigger');
     fireEvent.click(trigger);
@@ -77,7 +78,7 @@ describe('packages/mongo-nav/user-menu', () => {
 
   test('atlas MenuItem links to cloud.mongodb.com', () => {
     const { getByTestId } = renderUserMenu({
-      activeProduct: 'cloud',
+      activePlatform: 'cloud',
     });
     const trigger = getByTestId('user-menu-trigger');
     fireEvent.click(trigger);
@@ -92,7 +93,7 @@ describe('packages/mongo-nav/user-menu', () => {
 
   test('university MenuItem links to university.mongodb.com', () => {
     const { getByTestId } = renderUserMenu({
-      activeProduct: 'cloud',
+      activePlatform: 'cloud',
     });
     const trigger = getByTestId('user-menu-trigger');
     fireEvent.click(trigger);
@@ -107,7 +108,7 @@ describe('packages/mongo-nav/user-menu', () => {
 
   test('support MenuItem links to support.mongodb.com', () => {
     const { getByTestId } = renderUserMenu({
-      activeProduct: 'cloud',
+      activePlatform: 'cloud',
     });
     const trigger = getByTestId('user-menu-trigger');
     fireEvent.click(trigger);
@@ -122,7 +123,7 @@ describe('packages/mongo-nav/user-menu', () => {
 
   test('onLogout fires when logout is clicked', () => {
     const { getByText, getByTestId } = renderUserMenu({
-      activeProduct: 'cloud',
+      activePlatform: 'cloud',
     });
     const trigger = getByTestId('user-menu-trigger');
     fireEvent.click(trigger);
@@ -148,7 +149,7 @@ describe('packages/mongo-nav/user-menu', () => {
 
   test('does not render atlas MenuItems when a non-cloud product is active', () => {
     const { queryByText, getByTestId } = renderUserMenu({
-      activeProduct: 'account',
+      activePlatform: 'account',
     });
     const trigger = getByTestId('user-menu-trigger');
     fireEvent.click(trigger);
@@ -167,7 +168,7 @@ describe('packages/mongo-nav/user-menu', () => {
   describe('renders appropriate links to SubMenu Items based on overrides prop', () => {
     test('renders particular url override, when the urls prop is set', () => {
       const { getByTestId } = renderUserMenu({
-        activeProduct: 'university',
+        activePlatform: 'university',
       });
 
       const trigger = getByTestId('user-menu-trigger');